requirements.txt
psycopg2-binary==2.8.2
pyTelegramBotAPI==3.6.6
